% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{AvgAndReg}
\alias{AvgAndReg}
\title{Average mutation rate estimates and regularize variance estiamte}
\usage{
AvgAndReg(
  Muts_df,
  bg_pval,
  bg_rate,
  alpha_p,
  beta_p,
  slope_vect,
  int_vect,
  nreps,
  sig_T2
)
}
\arguments{
\item{Muts_df}{a dataframe}

\item{bg_pval}{Numeric; p-value cutoff for calling a nt's mutation rate as being no greater than the background error rate}

\item{bg_rate}{Numeric; background mutation rate assumed}

\item{alpha_p}{Numeric; shape1 of the beta prior used for mutation rate estimation}

\item{beta_p}{Numeric; shape2 of the beta prior used for mutation rate estimation}

\item{slope_vect}{Vector of variance trend slopes}

\item{int_vect}{Vector of variance trend intercepts}

\item{nreps}{Number of replicates}

\item{sig_T2}{Variance about variance trend}
}
\value{
dataframe
}
\description{
Average mutation rate estimates and regularize variance estiamte
}
