% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{DiffMutTest}
\alias{DiffMutTest}
\title{Test for differential mutation rates}
\usage{
DiffMutTest(Muts_df, lden, nreps, nu_o)
}
\arguments{
\item{Muts_df}{a dataframe}

\item{lden}{Denominator of test statistic}

\item{nreps}{Number of replicates}

\item{nu_o}{Prior degrees of freedom}
}
\value{
dataframe
}
\description{
Test for differential mutation rates
}
