% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{EstimateRates}
\alias{EstimateRates}
\title{Estimate mutation rates and variances}
\usage{
EstimateRates(Muts_df, alpha_p, beta_p)
}
\arguments{
\item{Muts_df}{a dataframe in form required by HDProbe}

\item{alpha_p}{Numeric; shape1 of the beta prior used for mutation rate estimation}

\item{beta_p}{Numeric; shape2 of the beta prior used for mutation rate estimation}
}
\value{
a dataframe of same form as Muts_df but with additional info
}
\description{
Estimate mutation rates and variances
}
