% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{FilterSites}
\alias{FilterSites}
\title{Filter out low read count sites for variance trend estimation}
\usage{
FilterSites(Muts_df, nreps, cutoff = 1000)
}
\arguments{
\item{Muts_df}{a dataframe in form required by HDProbe}
}
\value{
a dataframe of same form as Muts_df but with additional info
}
\description{
Filter out low read count sites for variance trend estimation
}
