% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{freqvar}
\alias{freqvar}
\title{Calculate variance with delta approximation using frequentist model}
\usage{
freqvar(nmut, n)
}
\arguments{
\item{nmut}{Numeric; number of mutations}

\item{n}{Numeric; number of reads}
}
\description{
Calculate variance with delta approximation using frequentist model
}
