% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{inv_var}
\alias{inv_var}
\title{Calculate variance with delta approximation}
\usage{
inv_var(mu, var)
}
\arguments{
\item{mu}{Numeric; expectation value of random variable}

\item{var}{Numeric; variance of random variable}
}
\description{
Calculate variance with delta approximation
}
