% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketidy.R
\name{maketidy}
\alias{maketidy}
\title{Put mutation probing data into tidy format}
\usage{
maketidy(
  wt_muts,
  ko_muts,
  nreps,
  target = FALSE,
  SHAPE = FALSE,
  totcut = 500,
  rate_cut = 4
)
}
\arguments{
\item{wt_muts}{Dataframe from csv containing mutation profiling data for reference sample}

\item{ko_muts}{Dataframe from csv containing mutation profiling data for experimental sample}

\item{nreps}{Numeric; number of replicates}

\item{target}{Logical; if TRUE, then data is treated as targeted probing data}

\item{SHAPE}{Logical; if TRUE, then data is treated as SHAPE data rather than DMS-seq data}

\item{totcut}{Numeric; read count cutoff to be considered for analysis}

\item{rate_cut}{Numeric; 1/(max mutation rate) allowed; used for SNP filtering}
}
\value{
a dataframe that can be passed to HDProbe
}
\description{
Put mutation probing data into tidy format
}
