% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{RepVarCalc}
\alias{RepVarCalc}
\title{Filter out low read count sites for variance trend estimation}
\usage{
RepVarCalc(Filter_df, nbin, alpha_p = 2, beta_p = 102)
}
\arguments{
\item{Filter_df}{a dataframe}

\item{nbin}{Number of read count bins}
}
\value{
a dataframe
}
\description{
Filter out low read count sites for variance trend estimation
}
