% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{sim_muts}
\alias{sim_muts}
\title{Simulate heteroskedastic mutational probing data}
\usage{
sim_muts(
  npos,
  nreps = 3,
  het_m = -0.7,
  het_b = 1.3,
  nsig = round(npos/10),
  mu_e = 0,
  sig_e = 0.75,
  lp_mean = -4.35,
  lp_sd = 0.8,
  r_mean = 6.75,
  r_sd = 1.2,
  sig_T = 0.05,
  phi = 20
)
}
\arguments{
\item{npos}{Numeric; number of positions to simulate mutational data for}

\item{nreps}{Numeric; number of replicates to simulate for each of 2 conditions}

\item{het_m}{Numeric; slope of sequencing depth-variance relationship on log-log scale}

\item{het_b}{Numeric; intercept of sequencing depth-variance relationship on log-log scale}

\item{nsig}{Numeric; number of significant changes in mutation rate}

\item{mu_e}{Numeric; differences in logit(mutation rate) for significant nts drawn from normal(mu_e, sig_e)}

\item{sig_e}{Numeric; effect size parameter (see mu_e for details)}

\item{lp_mean}{Numeric; average logit(mutation rate)}

\item{lp_sd}{Numeric; standard deviation of logit(mutation rates)}

\item{r_mean}{Numeric; Avg. read counts drawn from lognormal(r_mean, r_sd)}

\item{r_sd}{Numeric; read counts drawn from lognormal(r_mean, r_sd)}

\item{sig_T}{Numeric; standard deviation of logit(p) variances on log scale}

\item{phi}{Numeric; read counts for each sample drawn from negative binomial distribution with size = phi}
}
\value{
List with 2 componenets, one of which is a dataframe that can be passed to HDProbe
}
\description{
Simulate heteroskedastic mutational probing data
}
