% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{TrendVariance}
\alias{TrendVariance}
\title{Estimate variability about replicate variability trend}
\usage{
TrendVariance(Avg_df, estimate_var, slope_vect, int_vect, nreps)
}
\arguments{
\item{Avg_df}{a dataframe}

\item{slope_vect}{Vector of variance trend slopes}

\item{int_vect}{Vector of variance trend intercepts}

\item{nreps}{Number of replicates}

\item{estiamte_var}{Variance of log10(sample variance)}
}
\value{
a list containing a dataframe and a vector
}
\description{
Estimate variability about replicate variability trend
}
