% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{var_calc}
\alias{var_calc}
\title{Calculate variance with delta approximation using beta distribution}
\usage{
var_calc(alpha, beta)
}
\arguments{
\item{alpha}{Numeric; shape1 parameter of beta distribution}

\item{beta}{Numeric; shape2 parameter of beta distribution}
}
\description{
Calculate variance with delta approximation using beta distribution
}
