% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{HDProbe}
\alias{HDProbe}
\title{Perform differential mutation analysis}
\usage{
HDProbe(
  Muts_df,
  nreps,
  homosked = FALSE,
  bg_pval = 1,
  bg_rate = 0.002,
  alpha_p = 2,
  beta_p = 102
)
}
\arguments{
\item{Muts_df}{A dataframe in the form required by HDProbe}

\item{nreps}{Integer; Number of replicates}

\item{homosked}{Logical; if TRUE, then the slope of the mean-variance relationship is set to 0}

\item{bg_pval}{Numeric; p-value cutoff for calling a nt's mutation rate as being no greater than the background error rate}

\item{bg_rate}{Numeric; background mutation rate assumed}

\item{alpha_p}{Numeric; shape1 of the beta prior used for mutation rate estimation}

\item{beta_p}{Numeric; shape2 of the beta prior used for mutation rate estiamtion}
}
\value{
A list with three elements
}
\description{
Perform differential mutation analysis
}
