% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDProbe.R
\name{HDProbe}
\alias{HDProbe}
\title{Perform differential mutation analysis}
\usage{
HDProbe(
  Muts_df,
  nreps,
  homosked = FALSE,
  bg_pval = 1,
  bg_rate = 0.002,
  alpha_p = NULL,
  beta_p = NULL,
  filter_het = 1000,
  filter_hom = 100,
  One_ctl = FALSE,
  Gene_ctl = FALSE,
  var_of_var = NULL,
  ...
)
}
\arguments{
\item{Muts_df}{A dataframe in the form required by HDProbe}

\item{nreps}{Integer; Number of replicates}

\item{homosked}{Logical; if TRUE, then the slope of the mean-variance relationship is set to 0}

\item{bg_pval}{Numeric; p-value cutoff for calling a nt's mutation rate as being no greater than the background error rate}

\item{bg_rate}{Numeric; background mutation rate assumed}

\item{alpha_p}{Numeric; shape1 of the beta prior used for mutation rate estimation. If alpha_p or beta_p are NULL, then this is
estimated using beta distribution method of moments estimator}

\item{beta_p}{Numeric; shape2 of the beta prior used for mutation rate estimation. If alpha_p or beta_p are NULL, then this
is estimated using beta distribution method of moments estimator}

\item{One_ctl}{Logical; if TRUE, then mutation rate comparisons are made to the global average control sample mutation
rate}

\item{Gene_ctl}{Logical; if TRUE, then mutation rate comparisons are made to the
gene-wide average control sample mutation rate}

\item{var_of_var}{Variance of variance to be used to tune regularization. If NULL, this
is estimated empirically (i.e., from the data)}

\item{...}{Parameters that can be supplied to internal functions}
}
\value{
A list with three elements
}
\description{
Perform differential mutation analysis
}
